refactor(claude): promote screen composites into the hifi kit
Move byline, domainRow, poolFilters, the cycle slides, promiseSlide,
selCard/selRail, the offerCard family, offerRow, and a new teamstrip
builder into hifi/kit.ts; relocate the admin cockpit builders from
screens/admin.ts; drop the orphaned statTiles builder and its CSS.
Locked component-library contract appended to the components-tab brief.
Artifact build stays at zero warnings.
